Add SDL2 audio backend for PC
